Extract & DRY the newsletter subscription components
