+++ src/pages/backstage-weekly.js
@@ -4,9 +4,10 @@ import { createUseStyles } from 'react-jss';
 
 import { Lead, SEO, StickyFooter, InterstitialTitle } from 'components';
 import PostSummary from 'components/blog/PostSummary';
-import FormSubmissionModal from 'components/actions/FormSubmissionModal';
 import CallToAction from 'components/actions/NetlifyFormCallToAction';
 import HeadRssLink from 'components/blog/HeadRssLink';
+import { SubscribeToNewsletterSuccessModal } from 'components/actions/SubscribeToNewsletter';
+
 import { FORM_NAMES } from '../contactFormConstants';
 import roadieRLogo from '../../content/assets/roadie-r-764x764.png';
 
@@ -58,21 +59,10 @@ const BlogIndex = ({ data, location }) => {
 
       <HeadRssLink />
 
-      <FormSubmissionModal
+      <SubscribeToNewsletterSuccessModal
         modalOpen={modalOpen}
         handleCloseModal={handleCloseModal}
-        titleText="You're subscribed!"
-        bodyText={
-          <>
-            <p>We publish most Mondays so you&apos;ll receive your first edition soon.</p>
-            <p>
-              In the meantime, we&apos;d love to hear why you&apos;re excited about Backstage.
-              Please fill out this short survey...
-            </p>
-          </>
-        }
         siteMetadata={data.site.siteMetadata}
-        followOn="NEEDS_ANALYSIS_SURVEY"
         email={email}
       />
