+++ src/templates/BlogPost.js
@@ -2,25 +2,15 @@ import React, { useState } from 'react';
 import { graphql } from 'gatsby';
 import { createUseStyles } from 'react-jss';
 
-import { SEO, InterstitialTitle, Link, ContentHeader, StickyFooter } from 'components';
-import FormSubmissionModal from 'components/actions/FormSubmissionModal';
-import CallToAction from 'components/actions/NetlifyFormCallToAction';
+import { SEO, ContentHeader, StickyFooter } from 'components';
 import HeadRssLink from 'components/blog/HeadRssLink';
-import { FORM_NAMES } from '../contactFormConstants';
+import {
+  SubscribeToNewsletterSuccessModal,
+  SubscribeToNewsletterCTA,
+} from 'components/actions/SubscribeToNewsletter';
 
 const useStyles = createUseStyles((theme) => ({
   main: theme.preMadeStyles.content,
-
-  callToActionWrapper: {
-    paddingTop: 40,
-    paddingBottom: 40,
-    textAlign: 'center',
-  },
-
-  callToActionParagraph: {
-    marginBottom: 24,
-    marginTop: 0,
-  },
 }));
 
 const MAX_WIDTH_BREAKPOINT = 'md';
@@ -47,21 +37,10 @@ const BlogPostTemplate = ({ data, location }) => {
 
       <HeadRssLink />
 
-      <FormSubmissionModal
+      <SubscribeToNewsletterSuccessModal
         modalOpen={modalOpen}
         handleCloseModal={handleCloseModal}
-        titleText="You're subscribed!"
-        bodyText={
-          <>
-            <p>We publish most Mondays so you&apos;ll receive your first edition soon.</p>
-            <p>
-              In the meantime, we&apos;d love to hear why you&apos;re excited about Backstage.
-              Please fill out this short survey...
-            </p>
-          </>
-        }
         siteMetadata={data.site.siteMetadata}
-        followOn="NEEDS_ANALYSIS_SURVEY"
         email={email}
       />
 
@@ -73,23 +52,11 @@ const BlogPostTemplate = ({ data, location }) => {
           </article>
         </main>
 
-        <div className={classes.callToActionWrapper}>
-          <InterstitialTitle text="Become a Backstage expert" />
-
-          <p className={classes.callToActionParagraph}>
-            To get the latest news, deep dives into Backstage features, and a roundup of recent
-            open-source action, sign up for Roadie&apos;s Backstage Weekly.{' '}
-            <Link to="/backstage-weekly/">See recent editions.</Link>
-          </p>
-
-          <CallToAction
-            setModalOpen={setModalOpen}
-            buttonText="Subscribe"
-            netlifyFormName={FORM_NAMES.subscribeToNewsletter}
-            email={email}
-            setEmail={setEmail}
-          />
-        </div>
+        <SubscribeToNewsletterCTA
+          setModalOpen={setModalOpen}
+          email={email}
+          setEmail={setEmail}
+        />
       </StickyFooter>
     </>
   );
